Refactor modals into domain folders
