+++ src/frontend/src/components/modals/__tests__/ChangeZoneMethodModal.test.tsx
@@ -2,7 +2,8 @@
 import '@testing-library/jest-dom/vitest';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { act, cleanup, fireEvent, render, screen, waitFor, within } from '@testing-library/react';
-import { ModalHost, setStorageHandoffHandler } from '../ModalHost';
+import { ModalHost } from '../ModalHost';
+import { setStorageHandoffHandler } from '../zones/storageHandoff';
 import { useSimulationStore } from '@/store/simulation';
 import { useUIStore } from '@/store/ui';
 import { quickstartSnapshot } from '@/data/mockTelemetry';
