+++ src/frontend/src/components/modals/common/Feedback.tsx
@@ -1,0 +1,7 @@
+export interface FeedbackProps {
+  message: string;
+}
+
+export const Feedback = ({ message }: FeedbackProps) => (
+  <p className="text-sm text-warning">{message}</p>
+);
