+++ src/frontend/src/components/modals/common/index.ts
@@ -1,0 +1,2 @@
+export * from './Feedback';
+export * from './ActionFooter';
