+++ src/frontend/src/components/modals/session/ImportGameModal.tsx
@@ -1,0 +1,6 @@
+export const ImportGameModal = () => (
+  <p className="text-sm text-text-muted">
+    Import a JSON save exported from Weedbreed.AI. Files are validated before
+    <code>sim.restoreSnapshot</code> executes.
+  </p>
+);
