+++ src/frontend/src/components/modals/session/LoadGameModal.tsx
@@ -1,0 +1,6 @@
+export const LoadGameModal = () => (
+  <p className="text-sm text-text-muted">
+    Load slots will appear once the backend exposes deterministic save headers. Choose a slot to
+    dispatch <code>systemFacade.loadSave</code>.
+  </p>
+);
