+++ src/frontend/src/components/modals/session/NotificationsModal.tsx
@@ -1,0 +1,6 @@
+export const NotificationsModal = () => (
+  <p className="text-sm text-text-muted">
+    Notification center will group events from <code>sim.*</code>, <code>world.*</code>, and
+    <code>finance.*</code>. Streaming view arrives with live telemetry.
+  </p>
+);
