+++ src/frontend/src/components/modals/structures/CreateRoomModal.tsx
@@ -1,0 +1,153 @@
+import { useState } from 'react';
+import { ActionFooter, Feedback } from '@/components/modals/common';
+import type { SimulationBridge } from '@/facade/systemFacade';
+import { useSimulationStore } from '@/store/simulation';
+import { formatNumber } from '@/utils/formatNumber';
+
+export interface CreateRoomModalProps {
+  bridge: SimulationBridge;
+  closeModal: () => void;
+  context?: Record<string, unknown>;
+}
+
+export const CreateRoomModal = ({ bridge, closeModal, context }: CreateRoomModalProps) => {
+  const structureId = typeof context?.structureId === 'string' ? context.structureId : null;
+  const structure = useSimulationStore((state) =>
+    structureId
+      ? (state.snapshot?.structures.find((item) => item.id === structureId) ?? null)
+      : null,
+  );
+  const rooms = useSimulationStore((state) =>
+    structureId
+      ? (state.snapshot?.rooms.filter((room) => room.structureId === structureId) ?? [])
+      : [],
+  );
+  const [roomName, setRoomName] = useState('');
+  const [purpose, setPurpose] = useState('Grow Room');
+  const [area, setArea] = useState(50);
+  const [busy, setBusy] = useState(false);
+  const [feedback, setFeedback] = useState<string | null>(null);
+
+  if (!structure || !structureId) {
+    return (
+      <p className="text-sm text-text-muted">
+        Structure data unavailable. Select a structure to add room.
+      </p>
+    );
+  }
+
+  const handleCreate = async () => {
+    if (!roomName.trim()) {
+      setFeedback('Room name is required.');
+      return;
+    }
+    if (area > availableArea) {
+      setFeedback(
+        `Room area (${formatNumber(area)} m²) exceeds available space (${formatNumber(availableArea)} m²).`,
+      );
+      return;
+    }
+    setBusy(true);
+    setFeedback(null);
+    try {
+      const response = await bridge.sendIntent({
+        domain: 'world',
+        action: 'createRoom',
+        payload: {
+          structureId,
+          room: {
+            name: roomName.trim(),
+            purpose,
+            area,
+          },
+        },
+      });
+      if (!response.ok) {
+        const warning = response.errors?.[0]?.message ?? response.warnings?.[0];
+        setFeedback(warning ?? 'Room creation rejected by facade.');
+        return;
+      }
+      closeModal();
+    } catch (error) {
+      console.error('Failed to create room', error);
+      setFeedback('Connection error while creating room.');
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  const purposeOptions = [
+    { id: 'Grow Room', name: 'Grow Room', description: 'Cultivation and plant management' },
+    { id: 'Laboratory', name: 'Laboratory', description: 'Breeding and research activities' },
+    { id: 'Break Room', name: 'Break Room', description: 'Staff rest and recovery' },
+    { id: 'Sales Room', name: 'Sales Room', description: 'Commercial product sales' },
+  ];
+
+  const existingRoomArea = rooms.reduce((sum, room) => sum + room.area, 0);
+  const availableArea = Math.max(0, structure.footprint.area - existingRoomArea);
+
+  return (
+    <div className="grid gap-4">
+      <div className="grid gap-3">
+        <label className="grid gap-1 text-sm">
+          <span className="text-xs font-semibold uppercase tracking-wide text-text-muted">
+            Room name
+          </span>
+          <input
+            type="text"
+            value={roomName}
+            onChange={(event) => setRoomName(event.target.value)}
+            placeholder="Enter room name"
+            className="w-full rounded-lg border border-border/60 bg-surface-muted/50 px-3 py-2 text-sm text-text focus:border-primary focus:outline-none"
+          />
+        </label>
+        <label className="grid gap-1 text-sm">
+          <span className="text-xs font-semibold uppercase tracking-wide text-text-muted">
+            Purpose
+          </span>
+          <select
+            value={purpose}
+            onChange={(event) => setPurpose(event.target.value)}
+            className="w-full rounded-lg border border-border/60 bg-surface-muted/50 px-3 py-2 text-sm text-text focus:border-primary focus:outline-none"
+          >
+            {purposeOptions.map((option) => (
+              <option key={option.id} value={option.id}>
+                {option.name} - {option.description}
+              </option>
+            ))}
+          </select>
+        </label>
+        <label className="grid gap-1 text-sm">
+          <span className="text-xs font-semibold uppercase tracking-wide text-text-muted">
+            Area (m²)
+          </span>
+          <input
+            type="number"
+            value={area}
+            onChange={(event) => setArea(Number(event.target.value))}
+            min="1"
+            max={availableArea}
+            step="1"
+            className={`w-full rounded-lg border px-3 py-2 text-sm text-text focus:outline-none ${
+              area > availableArea
+                ? 'border-red-500 bg-red-50 focus:border-red-500'
+                : 'border-border/60 bg-surface-muted/50 focus:border-primary'
+            }`}
+          />
+          <span className={`text-xs ${area > availableArea ? 'text-red-600' : 'text-text-muted'}`}>
+            Available: {formatNumber(availableArea, { maximumFractionDigits: 0 })} m² (structure
+            footprint: {formatNumber(structure.footprint.area)} m²)
+          </span>
+        </label>
+      </div>
+      {feedback ? <Feedback message={feedback} /> : null}
+      <ActionFooter
+        onCancel={closeModal}
+        onConfirm={handleCreate}
+        confirmLabel={busy ? 'Creating…' : 'Create room'}
+        confirmDisabled={busy || area > availableArea}
+        cancelDisabled={busy}
+      />
+    </div>
+  );
+};
