+++ src/frontend/src/components/modals/structures/DeleteRoomModal.tsx
@@ -1,0 +1,5 @@
+export const DeleteRoomModal = () => (
+  <p className="text-sm text-text-muted">
+    Room deletion flow is not yet wired. Select a room and confirm via dashboard controls.
+  </p>
+);
