+++ src/frontend/src/components/modals/structures/DeleteStructureModal.tsx
@@ -1,0 +1,86 @@
+import { useState } from 'react';
+import { ActionFooter, Feedback } from '@/components/modals/common';
+import type { SimulationBridge } from '@/facade/systemFacade';
+import { useNavigationStore } from '@/store/navigation';
+import { useSimulationStore } from '@/store/simulation';
+
+export interface DeleteStructureModalProps {
+  bridge: SimulationBridge;
+  closeModal: () => void;
+  context?: Record<string, unknown>;
+}
+
+export const DeleteStructureModal = ({
+  bridge,
+  closeModal,
+  context,
+}: DeleteStructureModalProps) => {
+  const structureId = typeof context?.structureId === 'string' ? context.structureId : null;
+  const structure = useSimulationStore((state) =>
+    structureId
+      ? (state.snapshot?.structures.find((item) => item.id === structureId) ?? null)
+      : null,
+  );
+  const goToStructures = useNavigationStore((state) => state.goToStructures);
+  const [confirmation, setConfirmation] = useState('');
+  const [busy, setBusy] = useState(false);
+  const [feedback, setFeedback] = useState<string | null>(null);
+
+  if (!structure || !structureId) {
+    return (
+      <p className="text-sm text-text-muted">
+        Structure data unavailable. Select a structure to remove.
+      </p>
+    );
+  }
+
+  const handleDelete = async () => {
+    setBusy(true);
+    setFeedback(null);
+    try {
+      const response = await bridge.sendIntent({
+        domain: 'world',
+        action: 'deleteStructure',
+        payload: { structureId },
+      });
+      if (!response.ok) {
+        const warning = response.errors?.[0]?.message ?? response.warnings?.[0];
+        setFeedback(warning ?? 'Deletion rejected by facade.');
+        return;
+      }
+      goToStructures();
+      closeModal();
+    } catch (error) {
+      console.error('Failed to delete structure', error);
+      setFeedback('Connection error while deleting structure.');
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  const confirmMatches = confirmation.trim() === structure.name;
+
+  return (
+    <div className="grid gap-4">
+      <p className="text-sm text-text-muted">
+        Removing <span className="font-semibold text-text">{structure.name}</span> releases all
+        rooms and zones. Type the structure name to confirm.
+      </p>
+      <input
+        type="text"
+        value={confirmation}
+        onChange={(event) => setConfirmation(event.target.value)}
+        placeholder={structure.name}
+        className="w-full rounded-lg border border-border/60 bg-surface-muted/50 px-3 py-2 text-sm text-text focus:border-primary focus:outline-none"
+      />
+      {feedback ? <Feedback message={feedback} /> : null}
+      <ActionFooter
+        onCancel={closeModal}
+        onConfirm={handleDelete}
+        confirmLabel={busy ? 'Removing…' : 'Remove structure'}
+        confirmDisabled={!confirmMatches || busy}
+        cancelDisabled={busy}
+      />
+    </div>
+  );
+};
