+++ src/frontend/src/components/modals/structures/DuplicateRoomModal.tsx
@@ -1,0 +1,93 @@
+import { useState } from 'react';
+import { ActionFooter, Feedback } from '@/components/modals/common';
+import type { SimulationBridge } from '@/facade/systemFacade';
+import { useNavigationStore } from '@/store/navigation';
+import { useSimulationStore } from '@/store/simulation';
+
+export interface DuplicateRoomModalProps {
+  bridge: SimulationBridge;
+  closeModal: () => void;
+  context?: Record<string, unknown>;
+}
+
+export const DuplicateRoomModal = ({ bridge, closeModal, context }: DuplicateRoomModalProps) => {
+  const roomId = typeof context?.roomId === 'string' ? context.roomId : null;
+  const room = useSimulationStore((state) =>
+    roomId ? (state.snapshot?.rooms.find((item) => item.id === roomId) ?? null) : null,
+  );
+  const zones = useSimulationStore((state) =>
+    roomId ? (state.snapshot?.zones.filter((zone) => zone.roomId === roomId) ?? []) : [],
+  );
+  const openRoom = useNavigationStore((state) => state.openRoom);
+  const [nameOverride, setNameOverride] = useState('');
+  const [busy, setBusy] = useState(false);
+  const [feedback, setFeedback] = useState<string | null>(null);
+
+  if (!room || !roomId) {
+    return (
+      <p className="text-sm text-text-muted">Room data unavailable. Select a room to duplicate.</p>
+    );
+  }
+
+  const handleDuplicate = async () => {
+    setBusy(true);
+    setFeedback(null);
+    try {
+      const payload: Record<string, unknown> = { roomId };
+      if (nameOverride.trim()) {
+        payload.name = nameOverride.trim();
+      }
+      const response = await bridge.sendIntent<{ roomId?: string }>({
+        domain: 'world',
+        action: 'duplicateRoom',
+        payload,
+      });
+      if (!response.ok) {
+        const warning = response.errors?.[0]?.message ?? response.warnings?.[0];
+        setFeedback(warning ?? 'Duplication rejected by facade.');
+        return;
+      }
+      const newRoomId = response.data?.roomId;
+      if (newRoomId) {
+        openRoom(room.structureId, newRoomId);
+      }
+      closeModal();
+    } catch (error) {
+      console.error('Failed to duplicate room', error);
+      setFeedback('Connection error while duplicating room.');
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  return (
+    <div className="grid gap-4">
+      <div className="grid gap-1 text-sm text-text-muted">
+        <span className="font-medium text-text">{room.name}</span>
+        <span>
+          {zones.length} zones · {formatNumber(room.area)} m² · Purpose {room.purposeName}
+        </span>
+      </div>
+      <label className="grid gap-1 text-sm">
+        <span className="text-xs font-semibold uppercase tracking-wide text-text-muted">
+          Duplicate as
+        </span>
+        <input
+          type="text"
+          value={nameOverride}
+          onChange={(event) => setNameOverride(event.target.value)}
+          placeholder="Leave empty to auto-name"
+          className="w-full rounded-lg border border-border/60 bg-surface-muted/50 px-3 py-2 text-sm text-text focus:border-primary focus:outline-none"
+        />
+      </label>
+      {feedback ? <Feedback message={feedback} /> : null}
+      <ActionFooter
+        onCancel={closeModal}
+        onConfirm={handleDuplicate}
+        confirmLabel={busy ? 'Duplicating…' : 'Duplicate room'}
+        confirmDisabled={busy}
+        cancelDisabled={busy}
+      />
+    </div>
+  );
+};
