+++ src/frontend/src/components/modals/structures/DuplicateStructureModal.tsx
@@ -1,0 +1,113 @@
+import { useState } from 'react';
+import { ActionFooter, Feedback } from '@/components/modals/common';
+import type { SimulationBridge } from '@/facade/systemFacade';
+import { useNavigationStore } from '@/store/navigation';
+import { useSimulationStore } from '@/store/simulation';
+
+export interface DuplicateStructureModalProps {
+  bridge: SimulationBridge;
+  closeModal: () => void;
+  context?: Record<string, unknown>;
+}
+
+export const DuplicateStructureModal = ({
+  bridge,
+  closeModal,
+  context,
+}: DuplicateStructureModalProps) => {
+  const structureId = typeof context?.structureId === 'string' ? context.structureId : null;
+  const structure = useSimulationStore((state) =>
+    structureId
+      ? (state.snapshot?.structures.find((item) => item.id === structureId) ?? null)
+      : null,
+  );
+  const rooms = useSimulationStore((state) =>
+    structureId
+      ? (state.snapshot?.rooms.filter((room) => room.structureId === structureId) ?? [])
+      : [],
+  );
+  const zones = useSimulationStore((state) =>
+    structureId
+      ? (state.snapshot?.zones.filter((zone) => zone.structureId === structureId) ?? [])
+      : [],
+  );
+  const openStructure = useNavigationStore((state) => state.openStructure);
+  const [nameOverride, setNameOverride] = useState('');
+  const [busy, setBusy] = useState(false);
+  const [feedback, setFeedback] = useState<string | null>(null);
+
+  if (!structure || !structureId) {
+    return (
+      <p className="text-sm text-text-muted">
+        Structure data unavailable. Select a structure before duplicating.
+      </p>
+    );
+  }
+
+  const handleDuplicate = async () => {
+    setBusy(true);
+    setFeedback(null);
+    try {
+      const payload: Record<string, unknown> = { structureId };
+      if (nameOverride.trim()) {
+        payload.name = nameOverride.trim();
+      }
+      const response = await bridge.sendIntent<{ structureId?: string }>({
+        domain: 'world',
+        action: 'duplicateStructure',
+        payload,
+      });
+      if (!response.ok) {
+        const warning = response.errors?.[0]?.message ?? response.warnings?.[0];
+        setFeedback(warning ?? 'Duplication rejected by facade.');
+        return;
+      }
+      const newId = response.data?.structureId;
+      if (newId) {
+        openStructure(newId);
+      }
+      closeModal();
+    } catch (error) {
+      console.error('Failed to duplicate structure', error);
+      setFeedback('Connection error while duplicating structure.');
+    } finally {
+      setBusy(false);
+    }
+  };
+
+  return (
+    <div className="grid gap-4">
+      <div className="grid gap-1 text-sm text-text-muted">
+        <span className="font-medium text-text">{structure.name}</span>
+        <span>
+          {rooms.length} rooms · {zones.length} zones · Rent €
+          {formatNumber(structure.rentPerTick, { maximumFractionDigits: 0 })} per tick
+        </span>
+      </div>
+      <label className="grid gap-1 text-sm">
+        <span className="text-xs font-semibold uppercase tracking-wide text-text-muted">
+          Duplicate as
+        </span>
+        <input
+          type="text"
+          value={nameOverride}
+          onChange={(event) => setNameOverride(event.target.value)}
+          placeholder="Leave empty to keep generated name"
+          className="w-full rounded-lg border border-border/60 bg-surface-muted/50 px-3 py-2 text-sm text-text focus:border-primary focus:outline-none"
+        />
+      </label>
+      <p className="text-xs text-text-muted">
+        The facade replicates geometry, rooms, zones, and device placement. Costs are applied on
+        commit; no optimistic updates are rendered.
+      </p>
+      {feedback ? <Feedback message={feedback} /> : null}
+      <ActionFooter
+        onCancel={closeModal}
+        onConfirm={handleDuplicate}
+        confirmLabel={busy ? 'Duplicating…' : 'Duplicate structure'}
+        confirmDisabled={busy}
+        cancelDisabled={busy}
+      />
+    </div>
+  );
+};
